requirements.txt
Django==5.0.2
djangorestframework==3.14.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
plotly==5.18.0
joblib==1.3.2
python-dotenv==1.0.0
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-filter==23.5
django-widget-tweaks==1.5.0
tensorflow==2.15.0
opencv-python==4.9.0.80
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
requests==2.31.0
kaggle==1.6.12
django-storages==1.14.2
boto3==1.34.34
Django==5.0.2
djangorestframework==3.14.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.0
requests==2.31.0
kaggle==1.6.12
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34
django-cors-headers>=4.3.1
django-filter>=23.5 